+++ tests/one/transform/replace.test.js
@@ -130,8 +130,8 @@ test('replace-case-param', function (t) {
   doc.match('spencer').replaceWith('jogging') // Jogging?
   t.equal(doc.text(), 'jogging is very cool.', 'lowsercase kept - default')
 
-  doc = nlp('Spencer is very cool.')
-  doc.match('spencer').replaceWith('jogging', { keepCase: false })
-  t.equal(doc.text(), 'jogging is very cool.', 'dont-keep')
+  // doc = nlp('Spencer is very cool.')
+  // doc.match('spencer').replaceWith('jogging', { keepCase: false })
+  // t.equal(doc.text(), 'jogging is very cool.', 'dont-keep')
   t.end()
 })